afegir entities

+++ IberaDelivery/Models/Comment.cs
@@ -5,10 +5,10 @@ namespace IberaDelivery.Models
 {
     public partial class Comment
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string? Contens { get; set; }
-        public long UserId { get; set; }
-        public long ProductId { get; set; }
+        public int UserId { get; set; }
+        public int ProductId { get; set; }
 
         public virtual Product Product { get; set; } = null!;
         public virtual User User { get; set; } = null!;
